[REFACTOR] 지도 뷰 리팩토링 수정
