+++ Zipadoo/Zipadoo/Views/Home/PromiseDetailProgressBarView.swift
@@ -9,6 +9,7 @@ import SwiftUI
 import CoreLocation
 import MapKit
 
+/// PromiseDetailMapSubView에서 사용
 struct PromiseDetailProgressBarView: View {
     
     // 관찰 대상 객체로 사용되는 위치 정보 스토어
@@ -166,151 +167,6 @@ struct PromiseDetailProgressBarView: View {
                 .shadow(color: .primary, radius: 10, x: 5, y: 5)
             
         )
-    }
-}
-
-extension PromiseDetailProgressBarView {
-    /// 현재 거리 비율을 계산하여 반환하는 함수
-    func distanceRatio(depature: Double, arrival: Double) -> Double {
-        let current: Double = arrival - depature
-        var remainingDistance: Double = current / arrival
-        
-        // 만약 remainingDistance가 0보다 작다면 그냥 0을 반환
-        if remainingDistance < 0 {
-            remainingDistance = 0
-        }
-        // 현재거리/총거리 비율
-        return remainingDistance
-    }
-    
-    /// 두 지점 간의 직선 거리를 계산하는 함수
-    func straightDistance(x1: Double, y1: Double, x2: Double, y2: Double) -> Double {
-        let z: Double = sqrt(pow(x2 - x1, 2) + pow(y2 - y1, 2))
-        return z
-    }
-    
-    /// 현재 위치와 목적지 간의 비율을 계산하는 함수
-    func calculateRatio(location: Location) -> Double {
-        let totalDistance = straightDistance(x1: location.departureLatitude, y1: location.departureLongitude, x2: promise.latitude, y2: promise.longitude)
-        let remainingDistance = straightDistance(x1: location.currentLatitude, y1: location.currentLongitude, x2: promise.latitude, y2: promise.longitude)
-        return distanceRatio(depature: remainingDistance, arrival: totalDistance)
-    }
-    
-    /// 지구가 원이라고 가정할 때, 각도를 라디안으로 변환하는 함수
-    func degreesToRadians(_ degrees: Double) -> Double {
-        return degrees * .pi / 180.0
-    }
-    
-    /// 두 지점 간의 거리를 미터로 계산하는 함수
-    func calculateDistanceInMeters(x1: Double, y1: Double, x2: Double, y2: Double) -> Double {
-        let earthRadius = 6371000.0 // 지구의 반경 (미터)
-        
-        let lat1 = degreesToRadians(x1)
-        let lon1 = degreesToRadians(y1)
-        let lat2 = degreesToRadians(x2)
-        let lon2 = degreesToRadians(y2)
-        
-        let dLat = lat2 - lat1
-        let dLon = lon2 - lon1
-        
-        let a = sin(dLat/2) * sin(dLat/2) + cos(lat1) * cos(lat2) * sin(dLon/2) * sin(dLon/2)
-        let c = 2 * atan2(sqrt(a), sqrt(1-a))
-        
-        let distance = earthRadius * c
-        return distance
-    }
-    
-    /// 미터 단위의 거리를 형식화하여 문자열로 반환하는 함수
-    func formatDistance(_ distance: CLLocationDistance) -> String {
-        if distance < 1000 {
-            return String(format: "%.0f m", distance)
-        } else {
-            let distanceInKilometers = distance / 1000.0
-            return String(format: "%.2f km", distanceInKilometers)
-        }
-    }
-    
-    /// 약속 종료 여부를 확인하는 함수
-    func calculateTimeRemaining(targetTime: Double) -> Bool {
-        // 현재 시간을 가져오기
-        let currentTime = Date().timeIntervalSince1970
-        // 목표 시간과 현재 시간의 차이 계산
-        let timeDifference = targetTime - currentTime
-        
-        // 시간이 지났는지 여부를 반환
-        if timeDifference <= 0 {
-            // 약속 시간이 이미 지났음
-            return true
-        } else {
-            // 아직 약속 시간이 남아있음
-            return false
-        }
-    }
-}
-
-struct ProgressSubView: View {
-    let friends: LocationAndParticipant
-    @Binding var isShowingFriendSheet: Bool
-    @Binding var region: MapCameraPosition
-    let realRatio: Double
-    @State var ratio: Double = 0
-    @Binding var progressTrigger: Bool
-    @Binding var promiseFinishCheck: Bool
-    
-    // 바의 시작 위치를 결정하는 offsetX 계산
-    var offsetX: Double {
-        if ratio <= 0.3 {
-            // 0.3 이하일 때는 바의 시작 위치를 바끝에
-            return 14
-        } else if 0.7 <= ratio && ratio < 1 {
-            // 0.7 이상일 때는 바의 시작 위치를 더 안으로
-            return 42
-        } else {
-            // 그 외의 경우는 기본값 28
-            return 28
-        }
-    }
-    
-    var body: some View {
-        // 지도 위치를 움직이는 버튼
-        Button {
-            region = .region(MKCoordinateRegion(center: friends.location.currentCoordinate, latitudinalMeters: 1000, longitudinalMeters: 1000))
-            isShowingFriendSheet = false
-        } label: {
-            ZStack(alignment: .leading) {
-                // 전체 바의 배경
-                RoundedRectangle(cornerRadius: 10)
-                    .frame(height: 38)
-                    .foregroundColor(Color.gray)
-                
-                // 진행 상태를 나타내는 바
-                RoundedRectangle(cornerRadius: 10)
-                    .frame(width: CGFloat(ratio) * (UIScreen.main.bounds.width - 64), height: 38)
-                    .foregroundColor(Color.brown)
-                
-                // 약속 종료 여부와 진행 상태에 따라 두더지 이미지 및 위치 결정
-                if promiseFinishCheck && ratio >= 1 {
-                    Image(AuthStore.shared.currentUser?.moleDrillImageString ?? "doo1_1")
-                        .resizable()
-                        .frame(width: 38, height: 55.95)
-                        .foregroundColor(.green)
-                        .offset(x: CGFloat(ratio) * (UIScreen.main.bounds.width - 64) - CGFloat(offsetX) + 5)
-                } else {
-                    Image(AuthStore.shared.currentUser?.moleDrillImageString ?? "doo1_1")
-                        .resizable()
-                        .frame(width: 38, height: 55.95)
-                        .foregroundColor(.green)
-                        .offset(y: 15)
-                        .rotationEffect(.degrees(270))
-                        .offset(x: CGFloat(ratio) * (UIScreen.main.bounds.width - 64) - CGFloat(offsetX))
-                }
-            }
-        }
-        .task {
-            withAnimation(.easeIn) {
-                ratio = realRatio
-            }
-        }
     }
 }
